Added ability to create nested views

+++ jenkinsapi/views.py
@@ -1,8 +1,19 @@
+import logging
 import json
 from jenkinsapi.view import View
 from jenkinsapi.exceptions import UnknownView
 
+log = logging.getLogger(__name__)
+
 class Views(object):
+
+    # TODO @lechat 20130702: Add check that plugin for view actually exists in Jenkins
+    LIST_VIEW = 'hudson.model.ListView'
+    NESTED_VIEW = 'hudson.plugins.nested_view.NestedView'
+    MY_VIEW = 'hudson.model.MyView'
+    DASHBOARD_VIEW = 'hudson.plugins.view.dashboard.Dashboard'
+    PIPELINE_VIEW = 'au.com.centrumsystems.hudson.plugin.buildpipeline.BuildPipelineView'
+
     """
     An abstraction on a Jenkins object's views
     """
@@ -17,8 +28,9 @@ class Views(object):
         if view_name == 'All':
             raise ValueError('Cannot delete this view: %s' % view_name)
 
-        self[view_name].delete()
-        self.jenkins.poll
+        if view_name in self:
+            self[view_name].delete()
+            self.jenkins.poll()
 
     def __setitem__(self, name, value):
         raise NotImplementedError()
@@ -30,7 +42,8 @@ class Views(object):
                     row['url'],
                     row['name'],
                     self.jenkins)
-        raise UnknownView(view_name)
+        else:
+            return None
 
     def __iteritems__(self):
         """
@@ -68,27 +81,29 @@ class Views(object):
         self.jenkins.poll()
         return self
 
-    def create(self, str_view_name):
+    def create(self, view_name, view_type=LIST_VIEW):
         """
         Create a view
-        :param str_view_name: name of new view, str
+        :param view_name: name of new view, str
         :param person: Person name (to create personal view), str
         :return: new View obj or None if view was not created
         """
+        log.info('Creating "%s" view "%s"' % (view_type, view_name))
         #url = urlparse.urljoin(self.baseurl, "user/%s/my-views/" % person) if person else self.baseurl
-        try:
-            return self[str_view_name]
-        except KeyError:
-            pass
+
+        if view_name in self:
+            log.warn('View "%s" already exists' % view_name)
+            return self[view_name]
+
         url = '%s/createView' % self.jenkins.baseurl
         headers = {'Content-Type': 'application/x-www-form-urlencoded'}
         data = {
-            "name": str_view_name,
-            "mode": "hudson.model.ListView",
+            "name": view_name,
+            "mode": view_type,
             "Submit": "OK",
-            "json": json.dumps({"name": str_view_name, "mode": "hudson.model.ListView"})
+            "json": json.dumps({"name": view_name, "mode": view_type})
         }
 
         self.jenkins.requester.post_and_confirm_status(url, data=data, headers=headers)
         self.jenkins.poll()
-        return self[str_view_name]
+        return self[view_name]
